pyproject.toml
[tool.poetry]
name = "awsgannet-light"
version = "0.1.0"
description = ""
authors = ["Mogady <ahmedmadgyno@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
langchain = "0.0.281"
fastapi = "^0.103.0"
openai = "^0.28.0"
weaviate-client = "^3.23.2"
tiktoken = "^0.4.0"
boto3 = "^1.28.40"
pydantic-settings = "^2.0.3"
jq = "^1.5.0"
uvicorn = {extras = ["standard"], version = "^0.23.2"}


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
